new stream: Jeff and Dom Make a Game (#247)
I'm the eponymous Jeff, software engineer at Slack and alumnus of game studios WayForward and Zindagi. My friend Dom and I have a pair programming stream where we're building a game from scratch, and teaching each other how to do it in the process. We hope this is a good contribution to your list. Thank you!