Add C0d3Spaghetti (#265)
welcome to the list @heyAyushh sorry I didn't see the PR when I was on holiday.

I may edit the file and move you up into the middle of the list, but can't decide where you should go just right now. Alphabetically, I'm thinking you would be the first 'C' entry, But, reading it 1337 style, maybe somewhere in the middle. I'll sort it.

👍